requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# Data processing
polars>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0
ccxt>=3.0.0
websocket-client>=1.6.0
aiohttp>=3.8.5

# AI & ML
torch>=2.0.0
pytorch-lightning>=2.0.0
optuna>=3.0.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.15.0
dash>=2.10.0
matplotlib>=3.7.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Database
pymysql>=1, <2
sqlalchemy>=2.0.0 